fix: Moved RotationLoadContext to Helper folder.
