+++ RotationSolver/Data/LoadedAssembly.cs
@@ -8,5 +8,7 @@ namespace RotationSolver.Data
 {
     internal class LoadedAssembly
     {
+        public string Path { get; set; }
+        public string LastModified { get; set; }
     }
 }
