+++ RotationSolver/Helpers/RotationHelper.cs
@@ -26,78 +26,6 @@ internal static class RotationHelper
 {
     public static List<LoadedAssembly> LoadedCustomRotations = new();
 
-    private class RotationLoadContext : AssemblyLoadContext
-    {
-        readonly DirectoryInfo _directory;
-
-        static Dictionary<string, Assembly> _handledAssemblies;
-        public RotationLoadContext(DirectoryInfo directoryInfo) : base(true)
-        {
-            _directory = directoryInfo;
-        }
-
-
-
-        static RotationLoadContext()
-        {
-            var assemblies = new Assembly[]
-            {
-                typeof(RotationSolverPlugin).Assembly,
-                typeof(Resolver).Assembly,
-                typeof(DalamudPluginInterface).Assembly,
-                typeof(DataCenter).Assembly,
-                typeof(SheetAttribute).Assembly,
-                typeof(QuestDialogueText).Assembly,
-            };
-
-            _handledAssemblies = new Dictionary<string, Assembly>();
-
-            foreach (var assembly in assemblies)
-            {
-                _handledAssemblies.Add(assembly.GetName().Name, assembly);
-            }
-        }
-
-        protected override Assembly Load(AssemblyName assemblyName)
-        {
-            if (assemblyName.Name != null && _handledAssemblies.TryGetValue(assemblyName.Name, out Assembly value))
-            {
-                return value;
-            }
-
-            var file = Path.Join(_directory.FullName, $"{assemblyName.Name}.dll");
-            if (File.Exists(file))
-            {
-                try
-                {
-                    return LoadFromFile(file);
-                }
-                catch
-                {
-                    //
-                }
-            }
-            return base.Load(assemblyName);
-        }
-
-        internal Assembly LoadFromFile(string filePath)
-        {
-            using var file = File.Open(filePath, FileMode.Open, FileAccess.Read, FileShare.Read);
-            var pdbPath = Path.ChangeExtension(filePath, ".pdb");
-            if (!File.Exists(pdbPath))
-            {
-                PluginLog.Information($"Failed to load {pdbPath}");
-                return LoadFromStream(file);
-            }
-            using var pdbFile = File.Open(pdbPath, FileMode.Open, FileAccess.Read, FileShare.Read);
-            var assembly = LoadFromStream(file, pdbFile);
-
-            file.Dispose();
-            pdbFile.Dispose();
-            return assembly;
-        }
-    }
-
     public static readonly Dictionary<Assembly, AssemblyInfo> _assemblyInfos = new();
 
     public static string[] AllowedAssembly { get; private set; } = Array.Empty<string>();
