+++ RotationSolver/Updaters/RotationUpdater.cs
@@ -1,6 +1,6 @@
 ﻿using Dalamud.Logging;
 using Lumina.Excel.GeneratedSheets;
-
+using RotationSolver.Helpers;
 using RotationSolver.Localization;
 using System.Text;
